Add files via upload
This script extracts the residual water saturation in different slices or ROIs based on the signal intensity changes between a reference slices or ROIs and the corresponding target slices or ROIs. Reference slices must be in a folder called reference and the slice number must be equal to that in the target slices for proper indexing.